# openapi.yaml
openapi: 3.0.3
info:
  title: Alimenti API
  description: >
    API REST per la gestione di alimenti (composizione CREA/INRAN), tabelle
    di riferimento LARN, ricette, pazienti con piani alimentari e
    appuntamenti, piramide Smartfood. Trasformazione in API REST
    dell'applicativo "Registro Alimenti" (Project Work — Sviluppo di
    un'applicazione full-stack API-based per un'organizzazione del settore
    sanitario, CdS Informatica per le Aziende Digitali).

    Autenticazione: un singolo utente amministratore. Le operazioni di sola
    lettura su alimenti, LARN, ricette e smartfood sono pubbliche; le
    operazioni di scrittura su questi moduli e TUTTE le operazioni sui
    pazienti (dati clinici) richiedono il login (vedi /auth/login). La
    sessione e' mantenuta con un cookie firmato.
  version: "1.0.0"
servers:
  - url: /api
    description: Stesso host del front-end (modalita' consigliata)
...
security:
  - cookieAuth: []
paths:
...
  /pazienti:
    get:
      tags: [Pazienti]
      summary: Elenco paginato pazienti (richiede login)
      parameters:
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/page"
        - name: attivi
          in: query
          schema: { type: string, enum: ["0", "1"], default: "1" }
      responses:
        "200": { description: Elenco pazienti }
        "401": { $ref: "#/components/responses/NonAutenticato" }
    post:
      tags: [Pazienti]
      summary: Crea un paziente
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/PazienteInput" } }
      responses:
        "201":
          description: Paziente creato
          content:
            application/json: { schema: { $ref: "#/components/schemas/Paziente" } }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
  /pazienti/campi:
    get:
      tags: [Pazienti]
      summary: Metadati campi paziente/piano/giorno
      responses:
        "200": { description: Metadati }
        "401": { $ref: "#/components/responses/NonAutenticato" }
  /pazienti/{id}:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Pazienti]
      summary: Dettaglio paziente
      responses:
        "200":
          description: Paziente
          content:
            application/json: { schema: { $ref: "#/components/schemas/Paziente" } }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [Pazienti]
      summary: Aggiorna un paziente
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/PazienteInput" } }
      responses:
        "200": { description: Paziente aggiornato }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [Pazienti]
      summary: Elimina un paziente (CASCADE su piani e appuntamenti)
      responses:
        "204": { description: Eliminato }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /pazienti/{id}/larn:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Pazienti]
      summary: Fabbisogno di riferimento (energia/proteine/acqua) calcolato dalle tabelle LARN in base al profilo del paziente
      responses:
        "200": { description: Riferimenti LARN calcolati }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /pazienti/{id}/appuntamenti:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Pazienti]
      summary: Appuntamenti di un paziente
      responses:
        "200": { description: Elenco appuntamenti }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }

  # ------------------------------------------------------------------
  # Piani alimentari e giorni
  # ------------------------------------------------------------------
  /pazienti/{id}/piani:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Piani alimentari]
      summary: Elenco piani del paziente
      responses:
        "200": { description: Elenco piani }
        "401": { $ref: "#/components/responses/NonAutenticato" }
    post:
      tags: [Piani alimentari]
      summary: Crea un piano (genera automaticamente i 7 giorni della settimana)
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/PianoInput" } }
      responses:
        "201": { description: Piano creato, con i 7 giorni gia' predisposti }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
  /pazienti/{id}/piani/{piano_id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: piano_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Piani alimentari]
      summary: Dettaglio piano + panoramica dei 7 giorni
      responses:
        "200": { description: Piano e giorni }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [Piani alimentari]
      summary: Aggiorna l'intestazione del piano
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/PianoInput" } }
      responses:
        "200": { description: Piano aggiornato }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [Piani alimentari]
      summary: Elimina il piano (CASCADE sui 7 giorni)
      responses:
        "204": { description: Eliminato }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /pazienti/{id}/piani/{piano_id}/giorni/{giorno}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: piano_id
        in: path
        required: true
        schema: { type: integer }
      - $ref: "#/components/parameters/giornoPath"
    get:
      tags: [Piani alimentari]
      summary: Dettaglio di un giorno (pasti testuali + alimenti collegati + kcal/proteine)
      responses:
        "200": { description: Dettaglio giorno }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [Piani alimentari]
      summary: Aggiorna la descrizione testuale dei pasti del giorno
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object, description: "Uno o piu' campi pasto (es. colazione_latte, pranzo_primo, ...): vedi GET /pazienti/campi" } }
      responses:
        "200": { description: Giorno aggiornato }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /pazienti/{id}/piani/{piano_id}/giorni/{giorno}/ricalcola-kcal:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: piano_id
        in: path
        required: true
        schema: { type: integer }
      - $ref: "#/components/parameters/giornoPath"
    post:
      tags: [Piani alimentari]
      summary: Ricalcola e SALVA le kcal totali del giorno dagli alimenti collegati ai pasti
      responses:
        "200": { description: "kcal_stimate aggiornate" }
        "400": { description: Nessun alimento collegato }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /pazienti/{id}/piani/{piano_id}/giorni/{giorno}/alimenti:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: piano_id
        in: path
        required: true
        schema: { type: integer }
      - $ref: "#/components/parameters/giornoPath"
    post:
      tags: [Piani alimentari]
      summary: Collega un alimento (con quantita' in grammi) a un pasto del giorno
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/AlimentoPastoInput" } }
      responses:
        "201": { description: Alimento collegato }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /pazienti/{id}/piani/{piano_id}/giorni/{giorno}/alimenti/{link_id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: piano_id
        in: path
        required: true
        schema: { type: integer }
      - $ref: "#/components/parameters/giornoPath"
      - name: link_id
        in: path
        required: true
        schema: { type: integer }
    delete:
      tags: [Piani alimentari]
      summary: Rimuove un alimento collegato a un pasto
      responses:
        "204": { description: Rimosso }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }

  # ------------------------------------------------------------------
  # Appuntamenti
  # ------------------------------------------------------------------
  /appuntamenti:
    get:
      tags: [Appuntamenti]
      summary: Agenda appuntamenti su tutti i pazienti (con filtri)
      parameters:
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/page"
        - name: stato
          in: query
          schema: { type: string, enum: [Programmato, Completato, Annullato, "Non presentato"] }
        - name: passati
          in: query
          description: "1 per includere anche gli appuntamenti passati (default: solo futuri)"
          schema: { type: string, enum: ["0", "1"], default: "0" }
      responses:
        "200": { description: Agenda }
        "401": { $ref: "#/components/responses/NonAutenticato" }
    post:
      tags: [Appuntamenti]
      summary: Crea un appuntamento
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/AppuntamentoInput" } }
      responses:
        "201":
          description: Appuntamento creato
          content:
            application/json: { schema: { $ref: "#/components/schemas/Appuntamento" } }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
  /appuntamenti/campi:
    get:
      tags: [Appuntamenti]
      summary: Metadati campi + tipi/stati ammessi
      responses:
        "200": { description: Metadati }
        "401": { $ref: "#/components/responses/NonAutenticato" }
  /appuntamenti/{id}:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Appuntamenti]
      summary: Dettaglio appuntamento
      responses:
        "200":
          description: Appuntamento
          content:
            application/json: { schema: { $ref: "#/components/schemas/Appuntamento" } }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [Appuntamenti]
      summary: Aggiorna un appuntamento
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/AppuntamentoInput" } }
      responses:
        "200": { description: Appuntamento aggiornato }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [Appuntamenti]
      summary: Elimina un appuntamento
      responses:
        "204": { description: Eliminato }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie di sessione impostato da POST /auth/login.
  parameters:
    q:
      name: q
      in: query
      description: Ricerca libera
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
    giornoPath:
      name: giorno
      in: path
      required: true
      schema:
        type: string
        enum: [Lunedi, Martedi, Mercoledi, Giovedi, Venerdi, Sabato, Domenica]
  responses:
    NonTrovato:
      description: Risorsa non trovata
      content:
        application/json: { schema: { $ref: "#/components/schemas/Errore" } }
    NonAutenticato:
      description: Login richiesto
      content:
        application/json: { schema: { $ref: "#/components/schemas/Errore" } }
    DatiNonValidi:
      description: Dati non validi
      content:
        application/json: { schema: { $ref: "#/components/schemas/ErroreValidazione" } }
  schemas:
    Errore:
      type: object
      properties:
        error: { type: string }
    ErroreValidazione:
      type: object
      properties:
        error: { type: string }
        details:
          type: array
          items: { type: string }
...
    PazienteInput:
      type: object
      required: [nome, cognome]
      properties:
        nome: { type: string }
        cognome: { type: string }
        data_nascita: { type: string, format: date }
        sesso: { type: string, enum: [M, F] }
        altezza_cm: { type: number }
        peso_kg: { type: number }
        livello_attivita_fisica:
          type: string
          description: "Coefficiente PAL (es. 1.4, 1.6, 1.8, 2.0)"
        telefono: { type: string }
        email: { type: string }
        note_cliniche: { type: string }
        attivo: { type: boolean, default: true }
    Paziente:
      allOf:
        - type: object
          properties:
            id: { type: integer }
        - $ref: "#/components/schemas/PazienteInput"
    PianoInput:
      type: object
      properties:
        titolo: { type: string }
        data_inizio: { type: string, format: date }
        note: { type: string }
        attivo: { type: boolean, default: true }
    AlimentoPastoInput:
      type: object
      required: [pasto, codice_alimento, quantita_g]
      properties:
        pasto: { type: string, enum: [colazione, spuntino_mattina, pranzo, spuntino_pomeriggio, cena] }
        codice_alimento: { type: string }
        quantita_g: { type: number, exclusiveMinimum: 0 }
        note: { type: string }
    AppuntamentoInput:
      type: object
      required: [paziente_id, data_ora]
      properties:
        paziente_id: { type: integer }
        data_ora: { type: string, format: date-time }
        tipo: { type: string, description: 'es. "Prima visita", "Controllo"' }
        stato: { type: string, enum: [Programmato, Completato, Annullato, "Non presentato"], default: Programmato }
        note: { type: string }
    Appuntamento:
      allOf:
        - type: object
          properties:
            id: { type: integer }
            paziente_nome: { type: string }
            paziente_cognome: { type: string }
        - $ref: "#/components/schemas/AppuntamentoInput"